backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
python-multipart
matplotlib
groq